requirements.txt
redis
pytest
python-dotenv
pymongo
faiss-cpu
flake8
